build(shim): begin narrow-CK_ULONG (i686) support — widen ulong + CI job
ADR-0011 implementation step 1. On 32-bit-CK_ULONG targets (i686/armv7/Windows
x64) `cryptoki_sys::CK_ULONG` is u32, so reading a ulong attribute value and
storing it in CkAttributeValue::Ulong(u64) failed to compile. Widen with
`as u64` (the crate's documented cross-platform convention;
`#![allow(clippy::unnecessary_cast)]` already present).

Add a `narrow-client-i686` CI job (continue-on-error) that builds+tests the
client-side crates for i686-unknown-linux-gnu with gcc-multilib (for ring). It
is the surfacing mechanism for the remaining narrow-target compile gaps before
the width bridge lands; promote to a required gate once green.

+++ crates/shim/src/dispatch/general/helpers/mod.rs
@@ -3690,8 +3690,10 @@ unsafe fn ck_attrs_to_rust_result(
                 let v = unsafe { *(attr.pValue as *const CK_BBOOL) };
                 Some(CkAttributeValue::Bool(v != 0))
             } else if ck_type.is_ulong() && len == std::mem::size_of::<CK_ULONG>() {
+                // `CK_ULONG` is u32 on narrow (32-bit-CK_ULONG) targets; widen to
+                // the wire's u64 so the shim compiles on i686/armv7/Windows-x64.
                 let v = unsafe { *(attr.pValue as *const CK_ULONG) };
-                Some(CkAttributeValue::Ulong(v))
+                Some(CkAttributeValue::Ulong(v as u64))
             } else if len > MAX_SERIALIZABLE_BYTES {
                 return Err(CkRv::ARGUMENTS_BAD);
             } else {
